wip
* Upgrade to Spring Boot 3.5.7
* Bump to latest RCs
* Use Java 21
* Use Gradle 9.1.0
* Updates to match `fint-flyt-instance-gateway`
* Change namespace from `webinstance` to only `instance`
