+++ build.gradle.kts
@@ -1,15 +1,16 @@
 import org.springframework.boot.gradle.plugin.SpringBootPlugin
 
 plugins {
-    id("org.springframework.boot") version "3.4.5"
+    id("org.springframework.boot") version "3.5.7"
     id("io.spring.dependency-management") version "1.1.7"
-    kotlin("jvm") version "2.1.10"
-    kotlin("plugin.spring") version "2.1.10"
+    kotlin("jvm") version "2.2.21"
+    kotlin("plugin.spring") version "2.2.21"
     id("maven-publish")
-    id("org.jlleitschuh.gradle.ktlint") version "13.0.0"
+    id("org.jlleitschuh.gradle.ktlint") version "13.1.0"
+    id("com.github.ben-manes.versions") version "0.53.0"
 }
 
-group = "no.fintlabs"
+group = "no.novari"
 version = findProperty("version") ?: "1.0-SNAPSHOT"
 
 repositories {
@@ -36,24 +37,20 @@ dependencies {
     implementation("org.springframework.boot:spring-boot-starter-oauth2-client")
     implementation("org.springframework.boot:spring-boot-starter-oauth2-resource-server")
     implementation("org.springframework.boot:spring-boot-starter-validation")
-    implementation("org.springframework:spring-core:6.0.0")
-
-    api("org.apache.httpcomponents.client5:httpclient5:5.4.2")
+    implementation("org.springframework.kafka:spring-kafka")
 
     implementation("no.fint:fint-arkiv-resource-model-java:$apiVersion")
 
-    implementation("no.fintlabs:fint-flyt-web-resource-server:1.0.0-rc-1")
-
-    implementation("no.fintlabs:fint-flyt-cache:1.2.3")
-    implementation("org.springframework.kafka:spring-kafka")
-    implementation("no.fintlabs:fint-kafka:5.0.0-rc-8")
-    implementation("no.fintlabs:fint-flyt-kafka:1.0-SNAPSHOT")
+    implementation("no.novari:flyt-web-resource-server:2.0.0-rc-1")
+    implementation("no.novari:flyt-cache:2.0.0-rc-2")
+    implementation("no.novari:kafka:5.0.0-rc-19")
+    implementation("no.novari:flyt-kafka:4.0.0-rc-7")
 
     testImplementation(kotlin("test"))
     testImplementation("org.springframework.boot:spring-boot-starter-test")
     testImplementation("org.springframework.security:spring-security-test")
     testImplementation("com.ninja-squad:springmockk:4.0.2")
-    testImplementation("org.mockito.kotlin:mockito-kotlin:5.4.0")
+    testImplementation("org.mockito.kotlin:mockito-kotlin:6.1.0")
 }
 
 tasks.test {
@@ -76,17 +73,26 @@ java {
     withSourcesJar()
 }
 
-ktlint {
-    version.set("1.7.1")
-    ignoreFailures.set(false)
-    outputToConsole.set(true)
-    filter {
-        exclude("**/generated/**")
-    }
-}
-
 tasks.named("check") {
     dependsOn("ktlintCheck")
 }
 
-apply(from = "https://raw.githubusercontent.com/FINTLabs/fint-buildscripts/master/reposilite.ga.gradle")
+publishing {
+    repositories {
+        maven {
+            url = uri("https://repo.fintlabs.no/releases")
+            credentials {
+                username = System.getenv("REPOSILITE_USERNAME")
+                password = System.getenv("REPOSILITE_PASSWORD")
+            }
+            authentication {
+                create<BasicAuthentication>("basic")
+            }
+        }
+    }
+    publications {
+        create<MavenPublication>("maven") {
+            from(components["java"])
+        }
+    }
+}
