+++ settings.gradle.kts
@@ -1,4 +1,4 @@
 plugins {
     id("org.gradle.toolchains.foojay-resolver-convention") version "0.8.0"
 }
-rootProject.name = "fint-flyt-web-instance-gateway"
+rootProject.name = "flyt-web-instance-gateway"
