+++ src/main/kotlin/no/fintlabs/gateway/instance/exception/AbstractInstanceRejectedException.kt
@@ -1,0 +1,5 @@
+package no.fintlabs.gateway.instance.exception
+
+abstract class AbstractInstanceRejectedException(
+    message: String,
+) : RuntimeException(message)
