+++ src/main/kotlin/no/fintlabs/gateway/instance/model/ArchiveCaseIdRequestParams.kt
@@ -1,0 +1,6 @@
+package no.fintlabs.gateway.instance.model
+
+data class ArchiveCaseIdRequestParams(
+    val sourceApplicationId: Long,
+    val sourceApplicationInstanceId: String,
+)
