+++ src/main/kotlin/no/fintlabs/gateway/instance/validation/InstanceValidationException.kt
@@ -1,0 +1,5 @@
+package no.fintlabs.gateway.instance.validation
+
+class InstanceValidationException(
+    val validationErrors: List<InstanceValidationService.Error>,
+) : RuntimeException("Instance validation errors: $validationErrors")
