--- src/main/kotlin/no/fintlabs/gateway/webinstance/exception/AbstractInstanceRejectedException.kt
@@ -1,5 +1,0 @@
-package no.fintlabs.gateway.webinstance.exception
-
-abstract class AbstractInstanceRejectedException(
-    message: String,
-) : RuntimeException(message)
