--- src/main/kotlin/no/fintlabs/gateway/webinstance/exception/IntegrationDeactivatedException.kt
@@ -1,7 +1,0 @@
-package no.fintlabs.gateway.webinstance.exception
-
-import no.fintlabs.gateway.webinstance.model.Integration
-
-class IntegrationDeactivatedException(
-    val integration: Integration,
-) : RuntimeException("Integration is deactivated: $integration")
