--- src/main/kotlin/no/fintlabs/gateway/webinstance/model/SourceApplicationIdAndSourceApplicationIntegrationId.kt
@@ -1,6 +1,0 @@
-package no.fintlabs.gateway.webinstance.model
-
-data class SourceApplicationIdAndSourceApplicationIntegrationId(
-    val sourceApplicationId: Long,
-    val sourceApplicationIntegrationId: String,
-)
