--- src/main/kotlin/no/fintlabs/gateway/webinstance/validation/InstanceValidationException.kt
@@ -1,5 +1,0 @@
-package no.fintlabs.gateway.webinstance.validation
-
-class InstanceValidationException(
-    val validationErrors: List<InstanceValidationService.Error>,
-) : RuntimeException("Instance validation errors: $validationErrors")
